+++ src/Products/MeetingCPASLalouviere/config.py
@@ -14,6 +14,7 @@ CPASLALOUVIEREROLES['budgetimpactreviewers'] = 'MeetingBudgetImpactReviewer'
 CPASLALOUVIEREROLES['n1'] = 'MeetingN1'
 CPASLALOUVIEREROLES['n2'] = 'MeetingN2'
 CPASLALOUVIEREROLES['secretaire'] = 'MeetingSecretaire'
+CPASLALOUVIEREROLES['budgetimpactreviewers'] = 'MeetingBudgetImpactReviewer'
 PMconfig.MEETINGROLES.update(CPASLALOUVIEREROLES)
 
 STYLESHEETS = [{'id': 'meetingcpaslalouviere.css',
@@ -27,8 +28,9 @@ PMconfig.EXTRA_GROUP_SUFFIXES = [
 ]
 
 CPASLALOUVIEREMEETINGREVIEWERS = {
-    'meetingitemcpaslalouviere_workflow': OrderedDict([('reviewers', 'proposed_to_president'),
-                                            ('secretaire', 'proposed_to_secretaire'),
-                                            ('n2', 'proposed_to_n2'),
-                                            ('n1', 'proposed_to_n1'), ])}
+    'meetingitemcpaslalouviere_workflow': OrderedDict([('reviewers', ['proposed_to_president']),
+                                            ('secretaire', ['proposed_to_secretaire']),
+                                            ('n2', ['proposed_to_n2']),
+                                            ('n1', ['proposed_to_n1']),
+                                            ('budgetimpactreviewers', ['proposed_to_budgetimpact_reviewer']),])}
 PMconfig.MEETINGREVIEWERS.update(CPASLALOUVIEREMEETINGREVIEWERS)